refactor(github): split SummaryTab into reusable components
- Move ContributionTable to separate component
- Create StudentSelector component for student selection and ordering
- Extract StudentCard component for individual student items
- Create contributionUtils file for helper functions
- Simplify SummaryTab by using new components and utilities
